frontend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
web3==6.13.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
eventlet==0.33.3
dnspython==2.4.2
eth-account==0.9.0
eth-keys==0.4.0
eth-typing==3.5.2
eth-utils==2.3.1
hexbytes==0.3.1
